+++ error-prone-contrib/src/test/java/tech/picnic/errorprone/bugpatterns/RefasterCheckTest.java
@@ -42,6 +42,7 @@ public final class RefasterCheckTest {
           "BigDecimal",
           "Collection",
           "Comparator",
+          "DirectReturn",
           "DoubleStream",
           "Equality",
           "ImmutableList",
@@ -57,7 +58,6 @@ public final class RefasterCheckTest {
           "JUnit",
           "LongStream",
           "MapEntry",
-          "Method",
           "Mockito",
           "Multimap",
           "Null",
